Some refactoring and cleanup
