+++ src/hyperdrive-cli/commands/service/config/settings-hyperdrive.go
@@ -1,7 +1,6 @@
 package config
 
 import (
-	"github.com/gdamore/tcell/v2"
 	"github.com/nodeset-org/hyperdrive/shared/config/ids"
 )
 
@@ -43,25 +42,7 @@ func (configPage *HyperdriveConfigPage) createContent() {
 	layout := newStandardLayout()
 	configPage.layout = layout
 	layout.createForm(&masterConfig.Hyperdrive.Network, "Hyperdrive and TX Fee Settings")
-
-	// Return to the home page after pressing Escape
-	layout.form.SetInputCapture(func(event *tcell.EventKey) *tcell.EventKey {
-		if event.Key() == tcell.KeyEsc {
-			// Close all dropdowns and break if one was open
-			for _, param := range configPage.layout.parameters {
-				dropDown, ok := param.item.(*DropDown)
-				if ok && dropDown.open {
-					dropDown.CloseList(configPage.home.md.app)
-					return nil
-				}
-			}
-
-			// Return to the home page
-			configPage.home.md.setPage(configPage.home.homePage)
-			return nil
-		}
-		return event
-	})
+	layout.setupEscapeReturnHomeHandler(configPage.home.md, configPage.home.homePage)
 
 	// Set up the form items
 	formItems := createParameterizedFormItems(masterConfig.Hyperdrive.GetParameters(), layout.descriptionBox)
