+++ src/hyperdrive-cli/commands/service/config/settings-logging.go
@@ -1,7 +1,6 @@
 package config
 
 import (
-	"github.com/gdamore/tcell/v2"
 	"github.com/nodeset-org/hyperdrive/hyperdrive-cli/client"
 )
 
@@ -43,25 +42,7 @@ func (configPage *LoggingConfigPage) createContent() {
 	// Create the layout
 	configPage.layout = newStandardLayout()
 	configPage.layout.createForm(&configPage.masterConfig.Hyperdrive.Network, "Logging Settings")
-
-	// Return to the home page after pressing Escape
-	configPage.layout.form.SetInputCapture(func(event *tcell.EventKey) *tcell.EventKey {
-		if event.Key() == tcell.KeyEsc {
-			// Close all dropdowns and break if one was open
-			for _, param := range configPage.layout.parameters {
-				dropDown, ok := param.item.(*DropDown)
-				if ok && dropDown.open {
-					dropDown.CloseList(configPage.home.md.app)
-					return nil
-				}
-			}
-
-			// Return to the home page
-			configPage.home.md.setPage(configPage.home.homePage)
-			return nil
-		}
-		return event
-	})
+	configPage.layout.setupEscapeReturnHomeHandler(configPage.home.md, configPage.home.homePage)
 
 	// Set up the form items
 	configPage.loggingItems = createParameterizedFormItems(configPage.masterConfig.Hyperdrive.Logging.GetParameters(), configPage.layout.descriptionBox)
